moved the scorecard componenet to be called in the game card
